Update SSD in specifications
Put a different SSD in the ThinkCentre because the SSD that was in there before has been moved to a new computer because the ThinkCentre will soon no longer be my daily driver. I will still be using the ThinkCentre, which means that this configuration will still be updated, so there will be no concerns about that.