Add Major AppleALC (audio) Changes
This commit fixes nearly all of the audio issues that this ThinkCentre had, by adding a new layout-id that I have made, which fixes the issue of no microphone jack input. There are still a couple of (very) minor issues, so I will likely add those soon in a seperate document, with some workarounds on how to fix them. But, the issues also mostly happen in Windows, so I’m not sure if it’s just a hardware fault on my ThinkCentre? I’ll still add it as an issue just in case.

Also, AppleALC has now been even more debloated now, as I’ve compiled my own version. I may add a GitHub Action that automatically does this for me when a new AppleALC version comes out, but I’m not completely sure if I will yet.

Signed-off-by: UHDbits <uhdbits@duck.com>