Rename filename of "ocvalidate" action
This commit just renames "tests.yml" to "ocvalidate.yaml" so it's more clear exactly what it is. Also, I changed the file extension to ".yaml" because that just seems to be used more, and is technically the standard. The original action added in this repository also used ".yaml", so I renamed it to make it more consistent.