Start overhaul of issue/PR templates
This commit starts the overhaul of issue/PR (but mostly issue) templates. The original “issue template” has been turned into an “issue form” for bugs, which will hopefully make it easier for people to report bugs, and easier for me to solve the bugs. This is not finished yet, and I hope to add more issue templates for both questions, and possible enhancements.

I have made some slight changes to the pull request template, but they were just quickly made, and I will refine them soon.

I also changed all “.yaml” files to “.yml”, because although “.yaml” seems to be the official standard, GitHub suggests using “.yml”, and it just seems to be more popular in general.

Signed-off-by: UHDbits <uhdbits@duck.com>