Add very early version of revamped configuration
This commit adds an EXTREMELY early version (probably not-functional) version of a revamped configuration, that still has many changes to go. Here is a list of changes that have been made so far though. Again, note this configuration likely does not work, with Sonoma, nor Ventura or Monterey.

- Update OpenCore and all kexts to latest version (in the case of OpenCore, 0.9.2 to 0.9.3)
- Change SMBIOS to an actual Haswell one (iMac14,4). Support for Ventura+ will be added in a later commit.
- Remove SSDT-USBX.aml, as it is only needed for Skylake+ SMBIOS’.
- Reorder kexts in “config.plist”
- Add comments to all kexts, drivers, and tools in “config.plist”
- Set “CustomSMBIOSGuid” to true, and SMBIOS mode to Custom, so that Windows will see the ThinkCentre as a ThinkCentre instead of a Mac.
- Change LauncherOption to “Full”, so OpenCore will automatically add itself to the BIOS (this change may be reverted).
- Change Timeout to 10 seconds instead of 5
- Change certain Debug settings to debugging values, but since the settings only work on the DEBUG version of OpenCore, this shouldn’t change anything to the average user.
- Change “SecureBootModel” to x86legacy.
- Remove AMFI boot-args (will have to be added back later)
- Change APFS MinVersion and MinDate to Big Sur, since this version of the configuration doesn’t support Catalina or below anyways.

As you can tell, this is looking to be quite a large revamp of the configuration, and I’ve barely gotten started! I’m excited to continue and release this to the public, possibly before Sonoma is offically released.

Signed-off-by: UHDbits <uhdbits@duck.com>