Finalize new configuration
This commit should hopefully finalize the completely revamped configuration, with a couple of major final changes listed below.

- Add "AMFIPass.kext" so we can enable AMFI
- Enable AMFI and remove the other now unneeded boot argument
- Add "BrcmBluetoothInjector.kext" for Big Sur Bluetooth support
- Add "FeatureUnlock.kext" for a couple of features not supported on our SMBIOS.
- Do a bit of reorganization in the configuration
- Remove unneeded patches
- Change "MinKernel" for some of the patches
With this commit, this configuration should hopefully be mostly finalized other then OpenCore updates. I have plans for this major update to be the last major update to this configuration, as after this, this configuration will have nearly been perfected. I will still provide updates to the latest kexts and OpenCore versions, and also provide support and documentation updates, but other then that, I think it's time to change my focus to other configurations. Once this configuration is actually pushed to "main", I will make a more formal message. Thanks for the support over the past year of me making this OpenCore configuration.

Signed-off-by: UHDbits <uhdbits@duck.com>