Finalize last major release of configuration
Oh boy, this commit adds tons of documentation changes. I'll try to list them all below, but I apologize for anything that I may miss.

- Add better documentation for the M83 Tiny and the M93p computers.
- Add a "Maintenance.md", about the current maintenance status of this OpenCore configuration.
- Remove Sonoma references from "OCLP.md"
- Change "OCLP.md" to link to the AMFIPass version of OCLP.
- Update grammar in Security.md
- Update all the "About This Mac" photos, including changing the Monterey ones to Big Sur.
- Convert the "Missing Icons" images to sRGB.
- Add SSDTs for the M83 Tiny and M93p.
- Numerous documentation changes to the README.md

As mentioned, this will be the last major release of this configuration. If you would like more info on the future of this configuration, please read the notice at the top of the README.md.

(p.s. 0.9.4 is misleading, the configuration has not yet been updated)

Signed-off-by: UHDbits <uhdbits@duck.com>