Fix RTC wake issues
Recently, specifically with macOS Sonoma, I noticed that my keyboard and monitor would randomly light up during sleep, and then turn back off. Nothing would ever actually be output, but this was a sign the ThinkCentre was randomly waking up. After some diagnosis, it seems to be related to the RTC, and after adding this patch, everything seems to be fine.

Signed-off-by: UHDbits <uhdbits@duck.com>