Minor config.plist refinements
This commit mainly changes a couple of "Debug" settings in the config.plist, and also disables some Quirks that don't seem to be needed for this hardware. I've tested the configuration on my hardware, and everything seems to be working fine, but I decided to make this commit after the release of v0.9.7 in the hopes that if any of these changes cause issues, somebody else will report an issue with the "main" branch version of the configuration.

Signed-off-by: UHDbits <uhdbits@duck.com>